pyproject.toml
[project]
name = "archeologit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitpython>=3.1",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "streamlit>=1.54.0",
]
